:construction: TEST: Mapstruct problem 1.1

+++ src/main/java/com/thigas/quack/application/service/TaskService.java
@@ -2,7 +2,6 @@ package com.thigas.quack.application.service;
 
 import com.thigas.quack.adapter.dto.TaskDTO;
 import com.thigas.quack.adapter.mapper.CycleAvoidingMappingContext;
-import com.thigas.quack.adapter.mapper.TaskMapper;
 import com.thigas.quack.domain.repository.IStepRepository;
 import com.thigas.quack.domain.repository.ITaskRepository;
 import com.thigas.quack.infrastructure.persistence.entity.StepModel;
@@ -10,7 +9,9 @@ import com.thigas.quack.infrastructure.persistence.entity.TaskModel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -19,47 +20,47 @@ public class TaskService {
 
     @Autowired
     private ITaskRepository taskRepository;
+
     @Autowired
     private IStepRepository stepRepository;
 
     @Autowired
-    private TaskMapper taskMapper;
-
-    @Autowired
-    private CycleAvoidingMappingContext context;
-
+    private ObjectMapperService objectMapperService;
 
     public TaskDTO create(TaskDTO taskDTO) {
-        TaskModel taskModel = taskMapper.dtoToModel(taskDTO, context);
+        TaskModel taskModel = objectMapperService.toModel(taskDTO);
         Set<StepModel> stepEntities = new HashSet<>();
+
         for (Integer stepId : taskDTO.getSteps()) {
-            Optional<StepModel> task = stepRepository.findById(stepId);
-            task.ifPresent(stepEntities::add);
+            stepRepository.findById(stepId).ifPresent(stepEntities::add);
         }
+
         taskModel.setSteps(stepEntities);
         TaskModel savedTask = taskRepository.save(taskModel);
-        return taskMapper.modelToDto(savedTask, context);
+        return objectMapperService.toDto(savedTask);
     }
 
     public Optional<TaskDTO> getById(int id) {
-        Optional<TaskModel> taskOpt = taskRepository.findById(id);
-        return taskOpt.map(task -> taskMapper.modelToDto(task, new CycleAvoidingMappingContext()));
+        return taskRepository.findById(id)
+                .map(objectMapperService::toDto);
     }
 
     public Iterable<TaskDTO> getAll() {
         Iterable<TaskModel> tasks = taskRepository.findAll();
-        CycleAvoidingMappingContext context = new CycleAvoidingMappingContext();
         return StreamSupport.stream(tasks.spliterator(), false)
-                .map(task -> taskMapper.modelToDto(task, context))
+                .map(objectMapperService::toDto)
                 .collect(Collectors.toSet());
     }
 
     public void update(TaskDTO taskDTO) {
-        TaskModel task = taskMapper.dtoToModel(taskDTO, context);
-        taskRepository.save(task);
+        TaskModel taskModel = objectMapperService.toModel(taskDTO);
+        taskRepository.save(taskModel);
     }
 
     public void delete(int id) {
+        if (!taskRepository.existsById(id)) {
+            throw new IllegalArgumentException("Task não encontrada com id: " + id);
+        }
         taskRepository.deleteById(id);
     }
 }
