+++ src/main/java/com/thigas/quack/application/service/UserAchievementService.java
@@ -2,13 +2,12 @@ package com.thigas.quack.application.service;
 
 import com.thigas.quack.adapter.dto.UserAchievementDTO;
 import com.thigas.quack.adapter.mapper.CycleAvoidingMappingContext;
-import com.thigas.quack.adapter.mapper.StepMapper;
-import com.thigas.quack.adapter.mapper.UserAchievementMapper;
 import com.thigas.quack.domain.repository.IUserAchievementRepository;
 import com.thigas.quack.infrastructure.persistence.entity.UserAchievementModel;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -20,36 +19,39 @@ public class UserAchievementService {
     private IUserAchievementRepository userAchievementRepository;
 
     @Autowired
-    private UserAchievementMapper userAchievementMapper;
-
-    @Autowired
-    private CycleAvoidingMappingContext context;
-
+    private ObjectMapperService objectMapperService; // Usar ObjectMapperService
 
     public UserAchievementDTO create(UserAchievementDTO userAchievementDTO) {
-        UserAchievementModel userAchievementModel = userAchievementMapper.dtoToModel(userAchievementDTO, context);
+        UserAchievementModel userAchievementModel = objectMapperService.toModel(userAchievementDTO); // Usando ObjectMapperService
         UserAchievementModel savedUserModel = userAchievementRepository.save(userAchievementModel);
-        return userAchievementMapper.modelToDto(savedUserModel, context);
+        return objectMapperService.toDto(savedUserModel); // Usando ObjectMapperService
     }
 
     public Optional<UserAchievementDTO> getById(int id) {
-        Optional<UserAchievementModel> userAchievementOpt = userAchievementRepository.findById(id);
-        return userAchievementOpt.map(userAchievement -> userAchievementMapper.modelToDto(userAchievement, new CycleAvoidingMappingContext()));
+        return userAchievementRepository.findById(id)
+                .map(objectMapperService::toDto); // Usando ObjectMapperService
     }
 
     public Iterable<UserAchievementDTO> getAll() {
-        Iterable<UserAchievementModel> userAchievementsOpt = userAchievementRepository.findAll();
-        return StreamSupport.stream(userAchievementsOpt.spliterator(), false).map(userAchievement -> userAchievementMapper.modelToDto(userAchievement, new CycleAvoidingMappingContext()))
+        Iterable<UserAchievementModel> userAchievements = userAchievementRepository.findAll();
+        return StreamSupport.stream(userAchievements.spliterator(), false)
+                .map(objectMapperService::toDto) // Usando ObjectMapperService
                 .collect(Collectors.toList());
     }
 
     public void update(UserAchievementDTO userAchievementDTO) {
-        UserAchievementModel existingUser = userAchievementRepository.findById(userAchievementDTO.getId())
+        UserAchievementModel existingUserAchievement = userAchievementRepository.findById(userAchievementDTO.getId())
                 .orElseThrow(() -> new EntityNotFoundException("UserAchievement not found"));
-        userAchievementRepository.save(existingUser);
+
+        existingUserAchievement.setId(userAchievementDTO.getId());
+
+        userAchievementRepository.save(existingUserAchievement);
     }
 
     public void delete(int id) {
+        if (!userAchievementRepository.existsById(id)) {
+            throw new IllegalArgumentException("UserAchievement não encontrado com id: " + id);
+        }
         userAchievementRepository.deleteById(id);
     }
 }
