+++ src/main/java/com/thigas/quack/application/service/UserService.java
@@ -4,7 +4,6 @@ import com.thigas.quack.adapter.dto.UserDTO;
 import com.thigas.quack.adapter.mapper.UserMapper;
 import com.thigas.quack.domain.repository.IUserRepository;
 import com.thigas.quack.infrastructure.persistence.entity.UserModel;
-
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -30,19 +29,21 @@ public class UserService {
     }
 
     public Optional<UserDTO> getById(int id) {
-        Optional<UserModel> user = userRepository.findById(id);
-        return user.map(userMapper::modelToDto);
+        return userRepository.findById(id)
+                .map(userMapper::modelToDto);
     }
 
     public Iterable<UserDTO> getAll() {
         Iterable<UserModel> users = userRepository.findAll();
-        return StreamSupport.stream(users.spliterator(), false).map(userMapper::modelToDto)
+        return StreamSupport.stream(users.spliterator(), false)
+                .map(userMapper::modelToDto)
                 .collect(Collectors.toList());
     }
 
     public void update(UserDTO userDTO) {
         UserModel existingUser = userRepository.findById(userDTO.getId())
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
         if (userDTO.getName() != null) {
             existingUser.setName(userDTO.getName());
         }
@@ -54,13 +55,14 @@ public class UserService {
     }
 
     public void delete(int id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
 
-
     public Optional<UserDTO> findByEmail(String email) {
-        Optional<UserModel> user = userRepository.findByEmail(email);
-        return user.map(userMapper::modelToDto);
+        return userRepository.findByEmail(email)
+                .map(userMapper::modelToDto);
     }
-
 }
